read_influx_token_demo/requirements.txt
python-keycloak==0.18.0
requests==2.22.0